+++ common/pkg/testutils/capitest/request/items.go
@@ -15,8 +15,6 @@ import (
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 	capdv1 "sigs.k8s.io/cluster-api/test/infrastructure/docker/api/v1beta1"
 
-	capxv1 "github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/api/external/github.com/nutanix-cloud-native/cluster-api-provider-nutanix/api/v1beta1"
-	capav1 "github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/api/external/sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta2"
 	"github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/serializer"
 )
 
@@ -131,72 +129,6 @@ func NewKubeadmControlPlaneTemplateRequestItem(
 	return NewKubeadmControlPlaneTemplateRequest(uid, kubeadmControlPlaneTemplateRequestObjectName)
 }
 
-func NewAWSClusterTemplateRequestItem(
-	uid types.UID,
-	existingSpec ...capav1.AWSClusterTemplateSpec,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	awsClusterTemplate := &capav1.AWSClusterTemplate{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: capav1.GroupVersion.String(),
-			Kind:       "AWSClusterTemplate",
-		},
-	}
-
-	switch len(existingSpec) {
-	case 0:
-		// Do nothing.
-	case 1:
-		awsClusterTemplate.Spec = existingSpec[0]
-	default:
-		panic("can only take at most one existing spec")
-	}
-
-	return NewRequestItem(
-		awsClusterTemplate,
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			Kind:       "Cluster",
-			FieldPath:  "spec.infrastructureRef",
-			Name:       ClusterName,
-			Namespace:  Namespace,
-		},
-		uid,
-	)
-}
-
-func NewNutanixClusterTemplateRequestItem(
-	uid types.UID,
-	existingSpec ...capxv1.NutanixClusterTemplateSpec,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	nutanixClusterTemplate := &capxv1.NutanixClusterTemplate{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: capxv1.GroupVersion.String(),
-			Kind:       "NutanixClusterTemplate",
-		},
-	}
-
-	switch len(existingSpec) {
-	case 0:
-		// Do nothing.
-	case 1:
-		nutanixClusterTemplate.Spec = existingSpec[0]
-	default:
-		panic("can only take at most one existing spec")
-	}
-
-	return NewRequestItem(
-		nutanixClusterTemplate,
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			Kind:       "Cluster",
-			FieldPath:  "spec.infrastructureRef",
-			Name:       ClusterName,
-			Namespace:  Namespace,
-		},
-		uid,
-	)
-}
-
 func NewCPDockerMachineTemplateRequestItem(
 	uid types.UID,
 ) runtimehooksv1.GeneratePatchesRequestItem {
@@ -232,98 +164,6 @@ func NewWorkerDockerMachineTemplateRequestItem(
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      "docker-machine-template",
 				Namespace: "docker-cluster",
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			Kind:       "MachineDeployment",
-			FieldPath:  "spec.template.spec.infrastructureRef",
-		},
-		uid,
-	)
-}
-
-func NewCPAWSMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return NewRequestItem(
-		&capav1.AWSMachineTemplate{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: capav1.GroupVersion.String(),
-				Kind:       "AWSMachineTemplate",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "aws-machine-template",
-				Namespace: "aws-cluster",
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: controlplanev1.GroupVersion.String(),
-			Kind:       "KubeadmControlPlane",
-			FieldPath:  "spec.machineTemplate.infrastructureRef",
-		},
-		uid,
-	)
-}
-
-func NewWorkerAWSMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return NewRequestItem(
-		&capav1.AWSMachineTemplate{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: capav1.GroupVersion.String(),
-				Kind:       "AWSMachineTemplate",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "aws-machine-template",
-				Namespace: "aws-cluster",
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			Kind:       "MachineDeployment",
-			FieldPath:  "spec.template.spec.infrastructureRef",
-		},
-		uid,
-	)
-}
-
-func NewCPNutanixMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return NewRequestItem(
-		&capxv1.NutanixMachineTemplate{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: capxv1.GroupVersion.String(),
-				Kind:       "NutanixMachineTemplate",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "nutanix-machine-template",
-				Namespace: "nutanix-cluster",
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: controlplanev1.GroupVersion.String(),
-			Kind:       "KubeadmControlPlane",
-			FieldPath:  "spec.machineTemplate.infrastructureRef",
-		},
-		uid,
-	)
-}
-
-func NewWorkerNutanixMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return NewRequestItem(
-		&capxv1.NutanixMachineTemplate{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: capxv1.GroupVersion.String(),
-				Kind:       "NutanixMachineTemplate",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "nutanix-machine-template",
-				Namespace: "nutanix-cluster",
 			},
 		},
 		&runtimehooksv1.HolderReference{
